+++ src/Lockpick.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * Helps accessing protected/private members/consts of foreign objects
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
- * @copyright 2014-2023 Marcin Orlowski
+ * @copyright 2014-2024 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/php-lockpick
  */
@@ -21,14 +21,16 @@ class Lockpick
     /**
      * Calls protected method of $object, passing optional array of arguments.
      *
-     * @param object|string $clsOrObj Object to call $methodName on or name of the class.
-     * @param string $methodName Name of method to call.
-     * @param array $args Optional array of arguments (empty array for no args).
+     * @param object|string $clsOrObj   Object to call $methodName on or name of the class.
+     * @param string        $methodName Name of method to call.
+     * @param array         $args       Optional array of arguments (empty array for no args).
      *
-     * @throws \ReflectionException
-     * @throws \RuntimeException
+     * @throws \ReflectionException If method does not exist
+     * @throws \RuntimeException   If method is not accessible
      */
-    public static function call(string|object $clsOrObj, string $methodName, array $args = []): mixed
+    public static function call(string|object $clsOrObj,
+                                string        $methodName,
+                                array         $args = []): mixed
     {
         Validator::assertIsType($clsOrObj, [Type::EXISTING_CLASS, Type::OBJECT]);
 
@@ -44,13 +46,82 @@ class Lockpick
         return $method->invokeArgs(\is_object($clsOrObj) ? $clsOrObj : null, $args);
     }
 
-    /* **************************************************************************************************** */
+    /**
+     * Returns visibility of the method.
+     *
+     * @param object|string $clsOrObj   Class name to get method from, or instance of that class
+     * @param string        $methodName Method name to check
+     *
+     * @return Visibility Method visibility
+     * @throws \ReflectionException If method does not exist
+     */
+    public static function getMethodVisibility(object|string $clsOrObj,
+                                               string        $methodName): Visibility
+    {
+        $reflection = new \ReflectionClass($clsOrObj);
+        $method = $reflection->getMethod($methodName);
+
+        return $method->isPublic()
+            ? Visibility::PUBLIC
+            : ($method->isProtected()
+                ? Visibility::PROTECTED
+                : Visibility::PRIVATE);
+    }
+
+    /**
+     * Returns visibility of the class property.
+     *
+     * @param object|string $clsOrObj     Class name to get method from, or instance of that class
+     * @param string        $propertyName Property name to check
+     *
+     * @return Visibility Property visibility
+     * @throws \ReflectionException If property does not exist
+     */
+    public static function getPropertyVisibility(object|string $clsOrObj,
+                                                 string        $propertyName): Visibility
+    {
+        $reflection = new \ReflectionClass($clsOrObj);
+        $property = $reflection->getProperty($propertyName);
+
+        return $property->isPublic()
+            ? Visibility::PUBLIC
+            : ($property->isProtected()
+                ? Visibility::PROTECTED
+                : Visibility::PRIVATE);
+    }
+
+    /**
+     * Returns visibility of the class constant.
+     *
+     * @param object|string $clsOrObj     Class name to get method from, or instance of that class
+     * @param string        $constantName Constant name to check
+     *
+     * @return Visibility Constant visibility
+     * @throws \ReflectionException If consant does not exist
+     */
+    public static function getConstantVisibility(object|string $clsOrObj,
+                                                 string        $constantName): Visibility
+    {
+        $reflection = new \ReflectionClass($clsOrObj);
+        $constant = $reflection->getReflectionConstant($constantName);
+        if ($constant === false) {
+            throw new \ReflectionException("Failed obtaining constant: {$constantName}");
+        }
+
+        return $constant->isPublic()
+            ? Visibility::PUBLIC
+            : ($constant->isProtected()
+                ? Visibility::PROTECTED
+                : Visibility::PRIVATE);
+    }
+
+    /* ****************************************************************************************** */
 
     /**
      * Returns value of otherwise non-public property of the class
      *
      * @param string|object $clsOrObj Class name to get property from, or instance of that class
-     * @param string $name Property name to grab (i.e. `maxLength`)
+     * @param string        $name     Property name to grab (i.e. `maxLength`)
      *
      * @throws \ReflectionException
      */
@@ -87,13 +158,13 @@ class Lockpick
         $property->setValue($clsOrObj, $value);
     }
 
-    /* **************************************************************************************************** */
+    /* ****************************************************************************************** */
 
     /**
      * Returns value of otherwise non-public member of the class
      *
      * @param string|object $clsOrObj Class name to get member from, or instance of that class
-     * @param string $name Name of constant to grab (i.e. `FOO`)
+     * @param string        $name     Name of constant to grab (i.e. `FOO`)
      *
      * @return mixed
      *
@@ -111,6 +182,6 @@ class Lockpick
         return (new \ReflectionClass($clsOrObj))->getConstant($name);
     }
 
-    /* **************************************************************************************************** */
+    /* ****************************************************************************************** */
 
 }